requirements.txt
arxiv>=2.1.0
google-genai>=1.0.0
requests>=2.31.0
